Added CLI interface
