+++ .gitignore
@@ -56,3 +56,5 @@ typings/
 
 # dotenv environment variables file
 .env
+
+.idea/
